refactored star citizen code into separate proj
