+++ src/SSCM.Core/notes.md
@@ -23,8 +23,4 @@
 
 ## TODOs
 
-### Rename SCCM to ...? Space Sim Controls Manager
-
-### Refactor SC classes to separate library
-
 ### make SC a subcommand from the root for the CLI
