+++ src/SSCM.StarCitizen/ActionMapsXmlHelper.cs
@@ -3,7 +3,7 @@ using System.Xml;
 using System.Xml.Linq;
 using System.Xml.XPath;
 
-namespace SSCM.Core.SC;
+namespace SSCM.Core.StarCitizen;
 
 public class ActionMapsXmlHelper
 {
