+++ src/SSCM.Tests/Mocks/SCFoldersForTest.cs
@@ -1,4 +1,4 @@
-using SSCM.Core.SC;
+using SSCM.Core.StarCitizen;
 
 namespace SSCM.Tests.Mocks;
