+++ src/SSCM.Tests/SSCM.Tests.csproj
@@ -5,13 +5,14 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="NUnit" Version="3.13.3"/>
-    <PackageReference Include="NUnit.ConsoleRunner" Version="3.16.0"/>
-    <PackageReference Include="NUnit3TestAdapter" Version="4.3.1"/>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.1"/>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.2"/>
+    <PackageReference Include="NUnit" Version="3.13.3" />
+    <PackageReference Include="NUnit.ConsoleRunner" Version="3.16.0" />
+    <PackageReference Include="NUnit3TestAdapter" Version="4.3.1" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\SSCM.Core\SSCM.Core.csproj"/>
+    <ProjectReference Include="..\SSCM.Core\SSCM.Core.csproj" />
+    <ProjectReference Include="..\SSCM.StarCitizen\SSCM.StarCitizen.csproj" />
   </ItemGroup>
 </Project>
