+++ src/SSCM.cli/SSCM.cli.csproj
@@ -8,11 +8,12 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1"/>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.0"/>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0"/>
+    <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\SSCM.Core\SSCM.Core.csproj"/>
+    <ProjectReference Include="..\SSCM.Core\SSCM.Core.csproj" />
+    <ProjectReference Include="..\SSCM.StarCitizen\SSCM.StarCitizen.csproj" />
   </ItemGroup>
 </Project>
